add tip definition transformation
